backend/requirements.txt
flask
flask-cors
opencv-python-headless
requests
groq
pillow
pymysql
cryptography
gunicorn
python-dotenv
